コンポーネントのリファクタリング
