+++ __tests__/index.test.tsx
@@ -6,11 +6,11 @@
 
 import { render, screen } from '@testing-library/react'
 import React from 'react'
-import PrefecturesPopulation from '../pages/index'
+import Home from '../pages/index'
 
 describe('Title', () => {
   it('renders a heading', () => {
-    render(<PrefecturesPopulation />)
+    render(<Home />)
     const heading = screen.getByRole('heading', {
       name: /都道府県別人口構成/i,
     })
