+++ component/template/Layout.tsx
@@ -1,0 +1,22 @@
+import Head from 'next/head'
+import { ReactNode } from 'react'
+
+type LayoutProps = {
+  children: ReactNode
+  title: string
+}
+
+const Layout = ({ children, title }: LayoutProps) => {
+  return (
+    <>
+      <Head>
+        <title>{title}</title>
+        <meta name="description" content="prefectures population" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <main>{children}</main>
+    </>
+  )
+}
+
+export default Layout
